Create STUDENT GRADE CALCULATION
This program takes the number of subjects as input and then prompts the user to enter the marks obtained in each subject (out of 100). The total marks are calculated by adding up the marks obtained in all subjects. The average percentage is calculated by dividing the total marks by the total number of subjects and converting it to a percentage. The output is displayed to the console.