Create Number game
The optimized code calculates the user's score based on the number of attempts taken, with 10 points awarded for guessing the correct number on the first attempt, 9 points for guessing it on the second attempt, and so on. This makes the game more challenging and encourages the user to guess the correct number as quickly as possible. The code also uses a single sc.next() call to read the user's response to the "Do you want to play again?